requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
black==19.10b0
bson==0.5.9
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-gis==0.15
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
foursquare==1!2020.1.30
geographiclib==1.49
geopy==1.20.0
idna==2.9
importlib-metadata==0.17
isort==4.3.21
itsdangerous==1.1.0
Jinja2==3.0.0a1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==20.1
pathlib==1.0.1
pathspec==0.7.0
pep8==1.7.1
pluggy==0.12.0
psycopg2-binary==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
pytest==4.6.2
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.3.7
regex==2020.2.20
requests==2.23.0
six==1.12.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.8
wcwidth==0.1.7
Werkzeug==1.0.0
wrapt==1.11.2
zipp==0.5.1
